On peut afficher tous les étudiants

+++ src/main/java/gestionStages/dao/UserRepository.java
@@ -3,6 +3,6 @@ package gestionStages.dao;
 import gestionStages.entity.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Utilisateur, Long> {
+public interface UserRepository extends JpaRepository<Utilisateur, Integer> {
     Utilisateur findByUsername(String name);
 }
